feat: ERC721 Mintable, Burnable
